refactor: move the canvas card layers and spawners out of page.tsx
Refs #1936.

CanvasGlassPreviewPage moved from 5,463 lines to 4,753 lines.
